no more segfaults!
PyState_AddModule resulted in segfaults because the module wasn't
correctly initialized at the time it was called.

another public API, _PyImport_FixupExtensionObject, calls
_PyState_AddModule at a point when it makes sense to, and also tracks
import in a global dictionary.  so it's not the right API to use
*just* because it doesn't crash the interpreter.

but...

because we use imp to load all the modules, __package__ will not be
set the same way that it importlib sets it.  looks like we'll need to
have a separate code path for loading modules in python 3.